requirements.txt
langchain
langchain_community
langchain_huggingface
huggingface_hub
faiss-cpu
pypdf
streamlit
fastAPI
python-dotenv
langchain_google_genai